+++ JobSearchingSpringMVC/src/main/java/com/nhn/repository/impl/JobTypeRepositoryImpl.java
@@ -24,6 +24,12 @@ public class JobTypeRepositoryImpl implements JobTypeRepository {
     @Autowired
     private LocalSessionFactoryBean sessionFactory;
 
+    private final int maxItemsInPage = 10;
+
+    public int getMaxItemsInPage() {
+        return maxItemsInPage;
+    }
+
     @Override
     public JobType getById(int id) {
         Session session = this.sessionFactory.getObject().getCurrentSession();
@@ -61,7 +67,7 @@ public class JobTypeRepositoryImpl implements JobTypeRepository {
     }
 
     @Override
-    public List<JobType> getJobTypes(String name) {
+    public List<JobType> getJobTypes(String name, int page) {
         Session session = this.sessionFactory.getObject().getCurrentSession();
         CriteriaBuilder builder = session.getCriteriaBuilder();
         CriteriaQuery<JobType> query = builder.createQuery(JobType.class);
@@ -74,19 +80,29 @@ public class JobTypeRepositoryImpl implements JobTypeRepository {
         }
 
         Query q = session.createQuery(query);
+
+        if (page != 0) {
+            int max = maxItemsInPage;
+            q.setMaxResults(max);
+            q.setFirstResult((page - 1) * max);
+        }
         return q.getResultList();
     }
 
     @Override
     public boolean delete(JobType jobType) {
-        Session session = this.sessionFactory.getObject().getCurrentSession();
-        try {
-            session.delete(jobType);
-            return true;
-        } catch (HibernateException ex) {
-            System.err.println(ex.getMessage());
+        if (jobType.getJobPosts().size() == 0) {
+            Session session = this.sessionFactory.getObject().getCurrentSession();
+            try {
+                session.delete(jobType);
+                return true;
+            } catch (HibernateException ex) {
+                System.err.println(ex.getMessage());
+            }
+            return false;
+        } else {
+            return false;
         }
-        return false;
     }
 
     @Override
@@ -102,4 +118,13 @@ public class JobTypeRepositoryImpl implements JobTypeRepository {
 
         return false;
     }
+
+    @Override
+    public long count() {
+        Session session = this.sessionFactory.getObject().getCurrentSession();
+        Query q = session.createQuery("Select Count(*) From JobType");
+
+        return Long.parseLong(q.getSingleResult().toString());
+    }
+
 }
